third_party/blink/web_tests/platform/generic/fast/parser/document-write-onload-ordering-expected.txt

ALERT: 0
ALERT: 1
ALERT: 2
ALERT: 3
ALERT: 4
ALERT: 5
PASS
